+++ Config.lua
@@ -1256,6 +1256,7 @@ DF.PartyDefaults = {
     nameTextY = -10,
 
     -- Out of Range
+    oorAbsorbBarAlpha = 0.20000000298023,
     oorAurasAlpha = 0.20000000298023,
     oorBackgroundAlpha = 0.10000000149012,
     oorDefensiveIconAlpha = 0.5,
@@ -2344,6 +2345,7 @@ DF.RaidDefaults = {
     nameTextY = -10,
 
     -- Out of Range
+    oorAbsorbBarAlpha = 0.20000000298023,
     oorAurasAlpha = 0.20000000298023,
     oorBackgroundAlpha = 0.10000000149012,
     oorDefensiveIconAlpha = 0.5,